Add new variables to account for demand response
The new variables enable shifting demand from a time-slice to another time-slice (or other time-slices) within the same season. This shift is optimized by the model. There are three main constraints: 1) There might be cases of shifting the enitre demand of a commodity but generally we're allowed to shift only a fraction of a demand commodity. 2) Conservation of energy i.e. whatever petajoules from a demand commodity is not met in a time-slice, must be met some time during the day. The commodities for which demand response is feasible plus the fraction of possible demand shifting are specified in the "DemandResponse" parameter.  3) We can not put off meeting a demand too far. The maximum hours by which we're allowed to shift demand is specified by the parameter "MaxShiftingtime".