Add region_combination table and reformulate M.RegionalGlobalIndices rule
Previously, the M.RegionalGlobalIndices set was initialized by taking
all permutations of the regions. For models with 10+ regions, this
resulted in a set with 10e7+ members. This did not play well with
computer memory.

We now have introduced a separate table in the SQLite files, named
"regional_combinations" where modellers will list the region combinations
used in the model (for instance, an emissions constraint may be set over
a combination of regions, such as USA+CAN).

The syntax for this table is to separate the individual regions with a
'+' character.