Update ReserveMarginConstraint to allow for combination of regions
Previously, the region index only allowed for individual regions.
Now, the region index allows for combinations of regions (defined
in the "regions_combinations" table in addition to individual regions.

This allows users to pool regions together in reserve sharing zones.

Of note: the previous default for planning reserve margin was 0.2 (20%).
As a result of this commit, the default is changed to 0%.