Fix DemandActivityConstraint Bug
Previously, the demand activity constraint arbitrarily used the first timestep
in each period to determine the relative shares of end-use technologies.
This did not work as expected for demands for which there is zero demand in
the first timestep.

This commit fixes that by ensuring the timestep used to determine the relative shares
has a non-zero demand.